Move Swap CoinSelect to new package
