+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapx/coinselect/SelectSwapCoinViewModel.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.swap.coinselect
+package io.horizontalsystems.bankwallet.modules.swapx.coinselect
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
