+++ app/src/main/res/navigation/main_graph.xml
@@ -168,8 +168,7 @@
     <fragment
         android:id="@+id/swapFragment"
         android:name="io.horizontalsystems.bankwallet.modules.swapx.SwapXMainFragment"
-        android:label="@string/Swap"
-        tools:layout="@layout/fragment_swap"/>
+        android:label="@string/Swap" />
     <fragment
         android:id="@+id/swapApproveFragment"
         android:name="io.horizontalsystems.bankwallet.modules.swap.approve.SwapApproveFragment" />
@@ -302,7 +301,7 @@
         android:name="io.horizontalsystems.bankwallet.modules.info.TransactionStatusInfoFragment" />
     <fragment
         android:id="@+id/selectSwapCoinDialog"
-        android:name="io.horizontalsystems.bankwallet.modules.swap.coinselect.SelectSwapCoinFragment" />
+        android:name="io.horizontalsystems.bankwallet.modules.swapx.coinselect.SelectSwapCoinFragment" />
     <dialog
         android:id="@+id/wcErrorNoAccountFragment"
         android:name="io.horizontalsystems.bankwallet.modules.walletconnect.WCErrorNoAccountFragment" />
